# Optimize toCamelCase