Given

    >>> s = 'a_pretty_long_identifier_in_snake_case'

I get the following timeit results. Current implementation, with the @lru_cache decorator removed:

    >>> %timeit toCamelCase(s)
    10.4 µs ± 18.9 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

New version:

    >>> %timeit toCamelCase(s)
    4.88 µs ± 15.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

The LRU decorator is still there, since this is still a factor two below cache lookup speed:

    >>> %timeit toCamelCase(s)
    2.29 µs ± 21.3 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)