@gkoller 2021-03-11
```

    def test_toCamelCase():
        """Confirm that:
        - space separated name is converted to camelCase
        - PascalCase results in camelCase
        - snake_case results in camelCase
        """
        assert toCamelCase("test name magic") == "testNameMagic"
        assert toCamelCase("test name magic2") == "testNameMagic2"
        assert toCamelCase("testNameMagic") == "testNameMagic"
>       assert toCamelCase("TestNameMagic") == "testNameMagic"
E       AssertionError: assert 'TestNameMagic' == 'testNameMagic'
E         - testNameMagic
E         ? ^
E         + TestNameMagic
E         ? ^

tests/test_utils.py:13: AssertionError
```