@lbam 2021-03-11
Bug fixed at the expense of 0.4 µs per call.