@gkoller 2021-03-11
On my computer this is even faster (3x)r:
```python
CAMEL_CASE_REPLACE_REGEXP: Final[Pattern[str]] = re.compile(r"(?:_|\s)+(.)")

def toCamelCase(name: str) -> str:
    """
    Unify field/column/dataset name from Space separated/Snake Case/Camel case
    to camelCase.
    """
    result = CAMEL_CASE_REPLACE_REGEXP.sub(lambda mo: mo.group(1).upper(), name)
    return result[0].lower() + result[1:]
```