@lbam 2021-03-11
@gkoller I just discussed the new test with @ngaranko and we discovered that it was erroneous. It now checks for "numbers33InTheMiddle44" so your version is correct.