@lbam 2021-03-11
3x speedup confirmed, and that regexp isn't too horrible. I can get .1µs off by separating out the lambda expression. That gets us very close to LRU lookup speed.