@gkoller 2021-03-11
However my version will leave `numbers33inTheMiddle44`, when given as input, with a lower cased `in` in the output. :-(